Tests base*

+++ Tests/MainTests/TestDataModel.cs
@@ -2,17 +2,17 @@
 
 namespace MainTests
 {
-    public class TestDataModel : IDataModel<TestDataModel, long>
+    public class TestDataModel : IDataModel<TestDataModel, int>
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
         public string? Data { get; set; }
 
-        public static TestDataModel Factory(long id) => new()
+        public static TestDataModel Factory(int id) => new()
         {
             ID = id
         };
 
-        public bool EqualsByID(long id) => ID == id;
+        public bool EqualsByID(int id) => ID == id;
 
         public override string ToString() => $"[{ID}]: {Data ?? "null"}";
     }
